Add Dummy ExtractChildItems method for IconExtractor use
This commit introduces the `ExtractChildItems` method in the `BrowserItemFactory` class. This method uses the `ShellIconExtractor` to asynchronously extract icons for all child items in a specified folder. It includes parameters for the target folder and event handlers for icon extraction events. Additionally, a private method `IconExtOnComplete` is added to manage the completion of the extraction process. Comments and a reference to the source of the modifications are also included.